ai-ml-projects/image-processing/qr-code-tools/requirements.txt
qrcode[pil]==7.4.2
pyzbar==0.1.9
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.24.3
streamlit==1.31.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
